+++ wsidicom/file/io/frame_index/encapsulated_pixel_data.py
@@ -35,7 +35,7 @@ class EncapsulatedPixelData(FrameIndex):
         is empty.
 
         Returns
-        ----------
+        -------
         Optional[int]
             BOT length.
         """