+++ wsidicom/wsidicom.py
@@ -42,7 +42,7 @@ from wsidicom.geometry import Point, PointMm, Region, RegionMm, Size, SizeMm
 from wsidicom.graphical_annotations import AnnotationInstance
 from wsidicom.instance import WsiDataset, WsiInstance
 from wsidicom.optical import OpticalManager
-from wsidicom.series import Labels, Levels, Overviews
+from wsidicom.series import Labels, Overviews, Pyramid, Pyramids
 from wsidicom.source import Source
 from wsidicom.stringprinting import list_pretty_str
 from wsidicom.uid import SlideUids
@@ -70,9 +70,10 @@ class WsiDicom:
         source_owned: bool = False
             If source should be closed by this instance if used in a context manager.
         """
+        self._selected_pyramid = 0
         self._source = source
         self._source_owned = source_owned
-        self._levels = Levels.open(source.level_instances)
+        self._pyramids = Pyramids.open(source.level_instances)
         self._labels = Labels.open(source.label_instances)
         self._overviews = Overviews.open(source.overview_instances)
         self._annotations = list(source.annotation_instances)
@@ -99,7 +100,7 @@ class WsiDicom:
             streams for multiple files, or a path to a folder containing files.
 
         Returns
-        ----------
+        -------
         WsiDicom
             WsiDicom created from WSI DICOM files in path.
         """
@@ -134,7 +135,7 @@ class WsiDicom:
             supported transfer syntax is requested.
 
         Returns
-        ----------
+        -------
         WsiDicom
             WsiDicom created from WSI DICOM instances in study-series.
         """
@@ -165,7 +166,7 @@ class WsiDicom:
             Path to DICOMDIR file or directory with a DICOMDIR file.
 
         Returns
-        ----------
+        -------
         WsiDicom
             WsiDicom created from WSI DICOM files in DICOMDIR.
         """
@@ -186,7 +187,7 @@ class WsiDicom:
 
     def __repr__(self) -> str:
         return (
-            f"{type(self).__name__}({self.levels}, {self.labels}"
+            f"{type(self).__name__}({self.pyramids}, {self.labels}"
             f"{self.overviews}, {self.annotations})"
         )
 
@@ -195,36 +196,44 @@ class WsiDicom:
 
     @property
     def size(self) -> Size:
-        """Return pixel size of base level."""
-        return self.levels.size
+        """Return pixel size of base level of selected pyramid."""
+        return self.pyramids[self.selected_pyramid].size
 
     @property
     def mm_size(self) -> SizeMm:
-        return self.levels.mm_size
+        """Return image size in mm of selected pyramid."""
+        return self.pyramids[self.selected_pyramid].mm_size
 
     @property
     def tile_size(self) -> Size:
-        """Return tile size of levels."""
-        return self.levels.tile_size
+        """Return tile size of selected pyramid."""
+        return self.pyramids[self.selected_pyramid].tile_size
 
     @property
     def pixel_spacing(self) -> SizeMm:
-        return self.levels.pixel_spacing
+        """Return pixel spacing in mm/pixel of base level of selected pyramid."""
+        return self.pyramids[self.selected_pyramid].pixel_spacing
 
     @property
     def mpp(self) -> SizeMm:
-        return self.levels.mpp
+        """Return pixel size in um/pixel of base level of selected pyramid."""
+        return self.pyramids[self.selected_pyramid].mpp
+
+    @property
+    def selected_pyramid(self) -> int:
+        """Return index of selected pyramid."""
+        return self._selected_pyramid
 
     @property
     def uids(self) -> Optional[SlideUids]:
         return self._uids
 
     @property
-    def levels(self) -> Levels:
-        """Return contained levels."""
-        if self._levels is not None:
-            return self._levels
-        raise WsiDicomNotFoundError("levels", str(self))
+    def pyramids(self) -> Pyramids:
+        """Return contained pyramids."""
+        if self._pyramids is not None:
+            return self._pyramids
+        raise WsiDicomNotFoundError("pyramids", str(self))
 
     @property
     def labels(self) -> Optional[Labels]:
@@ -242,12 +251,12 @@ class WsiDicom:
         return self._annotations
 
     @property
-    def collection(self) -> List[Union[Levels, Labels, Overviews]]:
-        collection: List[Optional[Union[Levels, Labels, Overviews]]] = [
-            self._levels,
+    def collection(self) -> List[Union[Pyramid, Labels, Overviews]]:
+        collection: List[Optional[Union[Pyramid, Labels, Overviews]]] = [
             self._labels,
             self._overviews,
         ]
+        collection.extend(pyramid for pyramid in self._pyramids)
         return [series for series in collection if series is not None]
 
     def pretty_str(self, indent: int = 0, depth: Optional[int] = None) -> str:
@@ -258,9 +267,21 @@ class WsiDicom:
                 return string
         return (
             string
-            + " of levels:\n"
-            + list_pretty_str(self.levels.groups, indent, depth, 0, 2)
+            + " of pyramids:\n"
+            + list_pretty_str(list(self.pyramids), indent, depth, 0, 2)
         )
+
+    def set_selected_pyramid(self, index: int) -> None:
+        """Set selected pyramid.
+
+        Parameters
+        ----------
+        index: int
+            Index of pyramid to select.
+        """
+        if index < 0 or index >= len(self.pyramids):
+            raise WsiDicomNotFoundError(f"Pyramid {index}", str(self.pyramids))
+        self._selected_pyramid = index
 
     def read_label(self, index: int = 0) -> Image:
         """Read label image of the whole slide. If several label
@@ -272,17 +293,14 @@ class WsiDicom:
             Index of the label image to read.
 
         Returns
-        ----------
+        -------
         Image
-            label as image.
+            Label as Pillow image.
         """
         if self.labels is None:
             raise WsiDicomNotFoundError("label", str(self))
-        try:
-            label = self.labels[index]
-            return label.get_default_full()
-        except IndexError as exception:
-            raise WsiDicomNotFoundError("label", "series") from exception
+        label = self.labels.get(index)
+        return label.get_default_full()
 
     def read_overview(self, index: int = 0) -> Image:
         """Read overview image of the whole slide. If several overview
@@ -294,43 +312,48 @@ class WsiDicom:
             Index of the overview image to read.
 
         Returns
-        ----------
+        -------
         Image
-            Overview as image.
+            Overview as Pillow image.
         """
         if self.overviews is None:
             raise WsiDicomNotFoundError("overview", str(self))
-        try:
-            overview = self.overviews[index]
-            return overview.get_default_full()
-        except IndexError as exception:
-            raise WsiDicomNotFoundError("overview", "series") from exception
+        overview = self.overviews.get(index)
+        return overview.get_default_full()
 
     def read_thumbnail(
         self,
-        size: Tuple[int, int] = (512, 512),
+        size: Union[int, Tuple[int, int]] = 512,
         z: Optional[float] = None,
         path: Optional[str] = None,
+        pyramid: Optional[int] = None,
     ) -> Image:
-        """Read thumbnail image of the whole slide with dimensions
-        no larger than given size.
+        """Read thumbnail image of the whole slide with dimensions no larger than given
+        size.
 
         Parameters
         ----------
-        size: Tuple[int, int] = (512, 512)
+        size: Union[int, Tuple[int, int]] = 512
             Upper size limit for thumbnail.
         z: Optional[float] = None
             Z coordinate, optional.
         path: Optional[str] = None
             Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read thumbnail from. If `None` the index in `selected_pyramid` is
+            used.
 
         Returns
-        ----------
+        -------
         Image
-            Thumbnail as image,
+            Thumbnail as Pillow image,
         """
+        if isinstance(size, int):
+            size = (size, size)
         thumbnail_size = Size.from_tuple(size)
-        level = self.levels.get_closest_by_size(thumbnail_size)
+        if pyramid is None:
+            pyramid = self.selected_pyramid
+        level = self.pyramids.get(pyramid).get_closest_by_size(thumbnail_size)
         region = Region(position=Point(0, 0), size=level.size)
         image = level.get_region(region, z, path)
         image.thumbnail((size), resample=settings.pillow_resampling_filter)
@@ -343,6 +366,7 @@ class WsiDicom:
         size: Tuple[int, int],
         z: Optional[float] = None,
         path: Optional[str] = None,
+        pyramid: Optional[int] = None,
         threads: int = 1,
     ) -> Image:
         """Read region defined by pixels.
@@ -359,15 +383,20 @@ class WsiDicom:
             Z coordinate, optional.
         path: Optional[str] = None
             Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read region from. If `None` the index in `selected_pyramid` is
+            used.
         threads: int = 1
             Number of threads to use for read.
 
         Returns
-        ----------
+        -------
         Image
-            Region as image
+            Region as Pillow image.
         """
-        wsi_level = self.levels.get_closest_by_level(level)
+        if pyramid is None:
+            pyramid = self.selected_pyramid
+        wsi_level = self.pyramids.get(pyramid).get_closest_by_level(level)
         scale_factor = wsi_level.calculate_scale(level)
         scaled_region = (
             Region(position=Point.from_tuple(location), size=Size.from_tuple(size))
@@ -390,6 +419,7 @@ class WsiDicom:
         size: Tuple[float, float],
         z: Optional[float] = None,
         path: Optional[str] = None,
+        pyramid: Optional[int] = None,
         slide_origin: bool = False,
         threads: int = 1,
     ) -> Image:
@@ -398,27 +428,32 @@ class WsiDicom:
         Parameters
         ----------
         location: Tuple[float, float]
-            Upper left corner of region in mm, or lower left corner if using
-            slide origin.
+            Upper left corner of region in mm, or lower left corner if using slide
+            origin.
         level: int
-            Level in pyramid
+            Level in pyramid.
         size: Tuple[float, float].
-            Size of region in mm
+            Size of region in mm.
         z: Optional[float] = None
-            Z coordinate, optional
+            Z coordinate, optional.
         path: Optional[str] = None
-            optical path, optional
+            Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read region from. If `None` the index in `selected_pyramid` is
+            used.
         slide_origin: bool = False
             If to use the slide origin instead of image origin.
         threads: int = 1
             Number of threads to use for read.
 
         Returns
-        ----------
+        -------
         Image
-            Region as image
+            Region as Pillow image.
         """
-        wsi_level = self.levels.get_closest_by_level(level)
+        if pyramid is None:
+            pyramid = self.selected_pyramid
+        wsi_level = self.pyramids.get(pyramid).get_closest_by_level(level)
         scale_factor = wsi_level.calculate_scale(level)
         region = RegionMm(PointMm.from_tuple(location), SizeMm.from_tuple(size))
         image = wsi_level.get_region_mm(region, z, path, slide_origin, threads)
@@ -434,6 +469,7 @@ class WsiDicom:
         size: Tuple[float, float],
         z: Optional[float] = None,
         path: Optional[str] = None,
+        pyramid: Optional[int] = None,
         slide_origin: bool = False,
         threads: int = 1,
     ) -> Image:
@@ -442,8 +478,8 @@ class WsiDicom:
         Parameters
         ----------
         location: Tuple[float, float].
-            Upper left corner of region in mm, or lower left corner if using
-            slide origin.
+            Upper left corner of region in mm, or lower left corner if using slide
+            origin.
         mpp: float
             Requested pixel spacing (um/pixel).
         size: Tuple[float, float].
@@ -452,18 +488,23 @@ class WsiDicom:
             Z coordinate, optional.
         path: Optional[str] = None
             Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read region from. If `None` the index in `selected_pyramid` is
+            used.
         slide_origin: bool = False
             If to use the slide origin instead of image origin.
         threads: int = 1
             Number of threads to use for read.
 
         Returns
-        -----------
+        --------
         Image
-            Region as image
+            Region as Pillow image.
         """
         pixel_spacing = mpp / 1000.0
-        wsi_level = self.levels.get_closest_by_pixel_spacing(
+        if pyramid is None:
+            pyramid = self.selected_pyramid
+        wsi_level = self.pyramids.get(pyramid).get_closest_by_pixel_spacing(
             SizeMm(pixel_spacing, pixel_spacing)
         )
         region = RegionMm(PointMm.from_tuple(location), SizeMm.from_tuple(size))
@@ -479,32 +520,39 @@ class WsiDicom:
         tile: Tuple[int, int],
         z: Optional[float] = None,
         path: Optional[str] = None,
+        pyramid: Optional[int] = None,
     ) -> Image:
-        """Read tile in pyramid level as image.
+        """Read tile in pyramid level as Pillow image.
 
         Parameters
         ----------
         level: int
-            Pyramid level
+            Pyramid level.
         tile: int, int
-            tile xy coordinate
+            tile xy coordinate.
         z: Optional[float] = None
-            Z coordinate, optional
+            Z coordinate, optional.
         path: Optional[str] = None
-            optical path, optional
+            Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read tile from. If `None` the index in `selected_pyramid` is
+            used.
 
         Returns
-        ----------
+        -------
         Image
-            Tile as image
+            Tile as Pillow image.
         """
         tile_point = Point.from_tuple(tile)
+        wsi_pyramid = self.pyramids.get(
+            pyramid if pyramid is not None else self._selected_pyramid
+        )
         try:
-            wsi_level = self.levels.get_level(level)
+            wsi_level = wsi_pyramid.get(level)
             return wsi_level.get_tile(tile_point, z, path)
         except WsiDicomNotFoundError:
             # Scale from closest level
-            wsi_level = self.levels.get_closest_by_level(level)
+            wsi_level = wsi_pyramid.get_closest_by_level(level)
             return wsi_level.get_scaled_tile(tile_point, level, z, path)
 
     def read_encoded_tile(
@@ -513,6 +561,7 @@ class WsiDicom:
         tile: Tuple[int, int],
         z: Optional[float] = None,
         path: Optional[str] = None,
+        pyramid: Optional[int] = None,
     ) -> bytes:
         """Read tile in pyramid level as encoded bytes. For non-existing levels
         the tile is scaled down from a lower level, using the similar encoding.
@@ -520,30 +569,40 @@ class WsiDicom:
         Parameters
         ----------
         level: int
-            Pyramid level
+            Pyramid level.
         tile: int, int
-            tile xy coordinate
+            Tile xy coordinate.
         z: Optional[float] = None
-            Z coordinate, optional
+            Z coordinate, optional.
         path: Optional[str] = None
-            optical path, optional
+            Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read tile from. If `None` the index in `selected_pyramid` is
+            used.
 
         Returns
-        ----------
+        -------
         bytes
             Tile in file encoding.
         """
         tile_point = Point.from_tuple(tile)
+        wsi_pyramid = self.pyramids.get(
+            pyramid if pyramid is not None else self._selected_pyramid
+        )
         try:
-            wsi_level = self.levels.get_level(level)
+            wsi_level = wsi_pyramid.get(level)
             return wsi_level.get_encoded_tile(tile_point, z, path)
         except WsiDicomNotFoundError:
             # Scale from closest level
-            wsi_level = self.levels.get_closest_by_level(level)
+            wsi_level = wsi_pyramid.get_closest_by_level(level)
             return wsi_level.get_scaled_encoded_tile(tile_point, level, z, path)
 
     def get_instance(
-        self, level: int, z: Optional[float] = None, path: Optional[str] = None
+        self,
+        level: int,
+        z: Optional[float] = None,
+        path: Optional[str] = None,
+        pyramid: Optional[int] = None,
     ) -> WsiInstance:
         """Return instance fulfilling level, z and/or path.
 
@@ -552,16 +611,21 @@ class WsiDicom:
         level: int
             Pyramid level
         z: Optional[float] = None
-            Z coordinate, optional
+            Z coordinate, optional.
         path: Optional[str] = None
-            optical path, optional
+            Optical path, optional.
+        pyramid: Optional[int] = None
+            Pyramid to read tile from. If `None` the index in `selected_pyramid` is
+            used.
 
         Returns
-        ----------
+        -------
         WsiInstance:
             Instance
         """
-        wsi_level = self.levels.get_level(level)
+        if pyramid is None:
+            pyramid = self.selected_pyramid
+        wsi_level = self.pyramids.get(pyramid).get(level)
         return wsi_level.get_instance(z, path)
 
     def close(self) -> None:
@@ -576,6 +640,7 @@ class WsiDicom:
         workers: Optional[int] = None,
         chunk_size: Optional[int] = None,
         offset_table: Optional[Union["str", OffsetTableType]] = None,
+        include_pyramids: Optional[Sequence[int]] = None,
         include_levels: Optional[Sequence[int]] = None,
         include_labels: bool = True,
         include_overviews: bool = True,
@@ -607,8 +672,10 @@ class WsiDicom:
             `none`) or `OffsetTableType` enum. Default to None, which will use
             `bot` for encapsulated  syntaxes and `none` for non-encapsulated transfer
             syntaxes.
+        include_pyramids: Optional[Sequence[int]] = None
+            Optional list of indices (in present pyramids) to include.
         include_levels: Optional[Sequence[int]] = None
-            Optional list indices (in present levels) to include, e.g. [0, 1]
+            Optional list of indices (in all pyramids) to include, e.g. [0, 1]
             includes the two lowest levels. Negative indicies can be used,
             e.g. [-1, -2] includes the two highest levels.
         include_labels: bool = True
@@ -624,7 +691,7 @@ class WsiDicom:
             will be copied as is.
 
         Returns
-        ----------
+        -------
         List[Path]
             List of paths of created files.
         """
@@ -647,11 +714,12 @@ class WsiDicom:
             workers,
             chunk_size,
             offset_table,
+            include_pyramids,
             include_levels,
             add_missing_levels,
             transcoding,
         ) as target:
-            target.save_levels(self.levels)
+            target.save_pyramids(self.pyramids)
             if include_overviews and self.overviews is not None:
                 target.save_overviews(self.overviews)
             if include_labels:
@@ -676,7 +744,7 @@ class WsiDicom:
         Raises WsiDicomMatchError otherwise. Returns base uid for collection.
 
         Returns
-        ----------
+        -------
         SlideUids
             Matching uids
         """