+++ CHANGELOG.md
@@ -9,8 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for multiple pyramids within the same slide. A pyramid must have the same image coordinate system and extended depth of field (if any). Use the `pyramid`-parameter to set the pyramid in for example `read_region()`, or use `set_selected_pyramid()` to set the pyramid to use. By default the first detected pyramid is used.
 - RLE encoding using image codecs.
 - JPEG 2000 encoding of lossless YBR using image codecs.
+
+### Changed
+
+- Levels with different extended depth of fields are no longer considered to be the same pyramid.
 
 ## [0.17.0] - 2023-12-10